+++ helpful.el
@@ -1860,10 +1860,17 @@ imenu."
   ;; Prevent imenu converting "Source Code" to "Source.Code".
   (setq-local imenu-space-replacement " "))
 
+(defun helpful--flash-region (start end)
+  "Temporarily highlight region from START to END."
+  (let ((overlay (make-overlay start end)))
+    (overlay-put overlay 'face 'highlight)
+    (run-with-timer 1.0 nil 'delete-overlay overlay)))
+
 (defun helpful-visit-reference ()
   "Go to the reference at point."
   (interactive)
-  (let* ((path (get-text-property (point) 'helpful-path))
+  (let* ((sym helpful--sym)
+         (path (get-text-property (point) 'helpful-path))
          (pos (get-text-property (point) 'helpful-pos))
          (pos-is-start (get-text-property (point) 'helpful-pos-is-start)))
     (when (and path pos)
@@ -1889,7 +1896,13 @@ imenu."
       (when (or (< pos (point-min))
                 (> pos (point-max)))
         (widen))
-      (goto-char pos))))
+      (goto-char pos)
+      (save-excursion
+        (let ((defun-end (scan-sexps (point) 1)))
+          (while (re-search-forward
+                  (rx-to-string `(seq symbol-start ,(symbol-name sym) symbol-end))
+                  defun-end t)
+            (helpful--flash-region (match-beginning 0) (match-end 0))))))))
 
 (defun helpful-kill-buffers ()
   "Kill all `helpful-mode' buffers.